# livrables/openapi.json
{"openapi":"3.1.0","info":{"title":"Quiz API","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/question":{"get":{"tags":["Questions"],"summary":"Get All Questions","operationId":"get_all_questions_question_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Questions"],"summary":"Add Question","operationId":"add_question_question_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/question/{question_id}":{"get":{"tags":["Questions"],"summary":"Get Question By Id","operationId":"get_question_by_id_question__question_id__get","parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","title":"Question Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Questions"],"summary":"Update Question By Id","operationId":"update_question_by_id_question__question_id__put","parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","title":"Question Id"}},{"name":"question","in":"query","required":true,"schema":{"type":"string","title":"Question"}},{"name":"subject","in":"query","required":true,"schema":{"type":"string","title":"Subject"}},{"name":"use","in":"query","required":true,"schema":{"type":"string","title":"Use"}},{"name":"remark","in":"query","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remark"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_update_question_by_id_question__question_id__put"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Questions"],"summary":"Delete Question By Id","operationId":"delete_question_by_id_question__question_id__delete","parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","title":"Question Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"get":{"tags":["Users"],"summary":"Get Users","operationId":"get_users_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserModel"},"type":"array","title":"Response Get Users Users Get"}}}}}},"post":{"tags":["Users"],"summary":"Add User","operationId":"add_user_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"tags":["Users"],"summary":"Login User","operationId":"login_user_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"put":{"tags":["Users"],"summary":"Update User","operationId":"update_user_users__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateModel"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quizz/{quizz_id}":{"get":{"tags":["Quizz"],"summary":"Get Quizz By Id","operationId":"get_quizz_by_id_quizz__quizz_id__get","parameters":[{"name":"quizz_id","in":"path","required":true,"schema":{"type":"string","title":"Quizz Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Quizz"],"summary":"Delete Quizz By Id","operationId":"delete_quizz_by_id_quizz__quizz_id__delete","parameters":[{"name":"quizz_id","in":"path","required":true,"schema":{"type":"string","title":"Quizz Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quizz":{"get":{"tags":["Quizz"],"summary":"Get All Quizzs","operationId":"get_all_quizzs_quizz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/questions/{subject}":{"get":{"summary":"Get Questions","operationId":"get_questions_questions__subject__get","parameters":[{"name":"subject","in":"path","required":true,"schema":{"type":"string","title":"Subject"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_update_question_by_id_question__question_id__put":{"properties":{"correct":{"items":{"type":"string"},"type":"array","title":"Correct"},"responses":{"items":{"type":"string"},"type":"array","title":"Responses"},"good_answer_texte":{"items":{"type":"string"},"type":"array","title":"Good Answer Texte"}},"type":"object","required":["correct","responses","good_answer_texte"],"title":"Body_update_question_by_id_question__question_id__put"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QuestionModel":{"properties":{"question":{"type":"string","title":"Question"},"subject":{"type":"string","title":"Subject"},"use":{"type":"string","title":"Use"},"correct":{"items":{"type":"string"},"type":"array","title":"Correct"},"responses":{"items":{"type":"string"},"type":"array","title":"Responses"},"good_answer_texte":{"items":{"type":"string"},"type":"array","title":"Good Answer Texte"},"remark":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remark"}},"type":"object","required":["question","subject","use","correct","responses","good_answer_texte","remark"],"title":"QuestionModel"},"UserModel":{"properties":{"user_name":{"type":"string","title":"User Name"},"password_hash":{"type":"string","title":"Password Hash"},"role_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Role Id"}},"type":"object","required":["user_name","password_hash"],"title":"UserModel"},"UserUpdateModel":{"properties":{"user_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Name"},"password_hash":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password Hash"},"role_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Role Id"}},"type":"object","title":"UserUpdateModel"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}